requirements.txt
numpy
problog
pyswip
torchvision
torch
pillow
pysdd
pytest
setuptools